fix/use correct std for c or c++ with clangdlsp
There are still some bugs with this. For example,
if both a C and a C++ file are opened on different buffers,
the LSP tends to become confused and doesn't know what to do.
Perhaps the correct approach should be checking on init
of the LSP, instead of circumventing with a Vim function to get
the file extension name. For now, this works fine for basic use
of programming only in C++ or only in C at a time, but never
mixing them together in one Nvim instance.